requirements.txt
streamlit 
PyPDF2  
python-dotenv 
langchain 
langchain-community
torch
transformers 
faiss-cpu 